packages/core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "captchakraken"
version = "0.2.0"
description = "AI-powered captcha solver using attention-based coordinate extraction"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "CaptchaKraken Team", email = "info@example.com" },
]
dependencies = [
    "pydantic>=2.0.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "ollama>=0.3.0",
    "openai>=1.0.0",
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "accelerate>=0.27.0",
    "matplotlib>=3.7.0",
    "scipy>=1.10.0",
]

[project.scripts]
captchakraken = "captchakraken.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/captchakraken"]

